Extract generic checkout helpers into their own module
